pyproject.toml
[project]
name = "ansible-inv-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.25.1",
    "pyyaml>=6.0.3",
]
